refactor: use long long

+++ C++/PrimeNumber.cpp
@@ -2,14 +2,14 @@
 using namespace std;
 using ll = long long;
 
-vector<bool> prime_number(int n) {
+vector<bool> prime_number(ll n) {
   vector<bool> res(n + 1, true);
 
   res[0] = res[1] = false;
-  int m = ceil(sqrt(n));
-  for (int i = 2; i <= m; ++i) {
+  ll m = ceil(sqrt(n));
+  for (ll i = 2; i <= m; ++i) {
     if (res[i]) {
-      for (int j = i * 2; j <= n; j += i)
+      for (ll j = i * 2; j <= n; j += i)
         res[j] = false;
     }
   }
